requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
networkx>=3.1
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.8
soundfile>=0.11.0
pytest>=7.0.0